Created superclass for action types
