--- src/actionTypes/makeMdf/actions.js
@@ -1,7 +1,0 @@
-import makeMdfPcdc from './makeMdfPcdc.js';
-
-const actions = {
-  pcdc: makeMdfPcdc,
-};
-
-export default actions;
