--- src/actionTypes/validate/actions.js
@@ -1,7 +1,0 @@
-import validatePcdc from './validatePcdc.js';
-
-const actions = {
-  pcdc: validatePcdc,
-};
-
-export default actions;
